requirements.txt
pandas>=1.5.0
scikit-learn>=1.1.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
sqlalchemy>=1.4.0
pyarrow>=9.0.0
pyyaml>=6.0
requests>=2.28.0
pytest>=7.0.0
